pbkdf2: deprecate old `include_simple` API
This API is based on an MCF hash format (`$rpbkdf2$...`) which is unique
to this crate and therefore not interoperable.

As of #82, the PHC string format is supported, which provides an
interoperable way to serialize PBKDF2 password hashes.

It also implements the `McfHasher` trait which can be used to both
verify old `$rpbkdf2$` MCF hashes as well as upgrade them to PHC hashes.

This commit deprecates the old MCF API in the hope of getting people to
upgrade to the PHC format.