pbkdf2: remove legacy "simple" API
The previous "simple" API used an MCF hash format (`$rpbkdf2$...`) which
is unique to this crate and therefore not interoperable.

As of #82, the PHC string format is supported, which provides an
interoperable way to serialize PBKDF2 password hashes.

This commit completely replaces the old "simple" API with the
`password-hash` crate which supports the PHC string format as well as
an impl of the `McfHasher` trait which supports verifying and upgrading
hashes in the previous MCF format.