Bump password-hash to 0.1.2 due to breaking change
I had a use case where I was using the v0.7.3 of `pbkdf2`, and when dependabot bumped to v0.7.4 there was an error with compiling `pbkdf2`:

```
 if Base64::decode(count, &mut count_arr)?.len() != 4 {
    |                                                         ^ the trait `From<base64ct::Error>` is not implemented for `HasherError`
...
```
I found that this is due to my version of `password-hash` being stuck on v0.1.1. Once I did a `cargo update -p password-hash` all was resolved.