argon2: make `Params` always available; fix `hash_password_simple` (#182)
Closes #179.

This commit makes the `Params` struct always available regardless of
enabled crate features, adding on functionality when the `password-hash`
feature is enabled.

Per the added TODOs in this commit, it really seems like in the next
breaking release, `Argon2::new` should accept `Params` as an argument
rather than duplicating all of the individual parameter fields and
passing them in as arguments. After that, they can be validated and
stored within the `Argon2` struct itself. But that would be a breaking
change, so for now they're stored piecemeal.

Additionally, this commit now ensures enough context is available in the
`Argon2` struct in order to implement `hash_password_simple` that uses
the configured params rather than the defaults, which addresses the
aforementioned issue.