argon2: impl TryFrom<Params> for Argon2 (#202)
Support for initializing the algorithm from its parameters.

The `Params` struct has `pub` fields and therefore doesn't validate its
parameters, so we need a fallible conversion here.

An alternative would be to have `Params` always validate its parameters,
removing the `pub` fields and replacing them with reader methods. This
would allow an infallible `From` conversion instead.

This is noted as a TODO in params.rs.