password-auth: initial crate
Adds a high-level password hashing crate which is a facade over other
crates in this repo, namely the `argon2`, `pbkdf2`, and `scrypt` crates.

The crate has been designed with an extremely simple API with only
`generate_hash` and `verify_password` functions.

The API surface deliberately avoids exposing anything from other crates,
allowing for rapid iteration towards a 1.0 release even if the crates
it's built on are not yet 1.0.