Enable alloc feature of password-hash if enabled for argon2/balloon-hash
This commit enables the `password-hash/alloc` feature if the
`argon2/alloc` or `balloon-hash/alloc` features are enabled.

Prior to this commit, if a user adds argon2 as a dependency with the
`alloc` feature enabled, but not the `std` feature (which is the case
with the default feature set), then the `alloc` feature of
`password-hash` is not enabled. This means that feature-gated types such
as `PasswordHashString` are not available in the `password-hash`
re-export, so users wishing to use these types will have to add
`password-hash` as a separate dependency.