password-auth: add `is_hash_obsolete` (#428)
Adds a method for interrogating whether a given password hash is using
the recommended algorithm and parameters.

This can be used by password hash upgrade systems to determine whether a
particular hash should be upgraded/recomputed.

It additionally adds a `ParseError` type which wraps
`password_hash::Error`. This is only used by `is_hash_obsolete` for now
but it might be good for future versions to use this error type to
capture parse errors in `verify_password`.

However, that would be a breaking change, which is avoided for now.