argon2: fix big endian support
This was broken in #247, which added unsafe pointer casts to access the
contents of a block, which only works on little endian targets.

This reverts back to something closer to the previous code, which used
`u64::{from_le_bytes, to_le_bytes}` instead.

It also adds cross tests for PPC32 to spot future regressions.