argon2: add parallelism with cooperative views into memory
Coordinated shared access in the memory blocks is implemented with
`SegmentViewIter` and associated types, which provide views into Argon2
memory that can be processed in parallel.

These views alias in the regions that are read-only, but are disjoint in
the regions where mutation happens. Effectively, they implement, with a
combination of mutable borrowing and runtime checking, the cooperative
contract outlined in RFC 9106.

To avoid aliasing mutable references into the entire buffer of blocks
(which would be UB), pointers are used up to the moment where a
reference (shared or mutable) into a specific block is returned. At that
point, aliasing is no longer possible, as argued in SAFETY comments
and/or checked at runtime.

Finally, add a `parallel` feature and parallelize filling the blocks
using the memory views mentioned above and rayon.