argon2: add parallelism with cooperative views into memory
Coordinated shared access in the memory blocks is implemented with
`SegmentViewIter` and associated types, which provide views into the
Argon2 blocks memory that can be processed in parallel.

These views alias in the regions that are read-only, but are disjoint in
the regions where mutation happens. Effectively, they implement, with a
combination of mutable borrowing and runtime checking, the cooperative
contract outlined in RFC 9106.

To avoid aliasing mutable references into the entire buffer of blocks
(which would be UB), pointers are used up to the moment where a
reference (shared or mutable) into a specific block is returned. At that
point, aliasing is no longer possible, as argued in SAFETY comments
and/or checked at runtime.

The following tests have been executed in and pass Miri (modulo
unrelated warnings):

    reference_argon2i_v0x13_2_8_2
    reference_argon2id_v0x13_2_8_2

(Running these in Miri is quite slow, taking ~5 minutes each, so
only the most relevant tests were chosen).

Finally, add a default-enabled `parallel` feature, with a otherwise
optional dependency on `rayon`, and parallelize the filling of blocks
using the memory views mentioned above.