Add an example of peppering to the docs.
This patch adds an example of peppering as well as salting a
password to the library's documentation home page (right under the
existing example of salting). I think this is useful because
although the `new_with_secret()` method is public & documented,
it's purpose is not clear to the general user.