scrypt: make `Params` output len customization optional (#602)
This more or less goes back to the original API prior to when #255 added
output length customization, adding a separate constructor function for
adding length customization.

This keeps the output length out-of-the-way for users of the low-level
`scrypt::scrypt` API. We've had several complaints about this (#415,
#449, #596, #601), and hopefully this addresses them.

Additionally, since this functionality is only relevant to users of the
`password-hash` API, it's feature-gated on that as well.