scrypt: fixup doc compilation (#606)
When scrypt used a dependency with `default-features = false`, the doc
compilation will fail because `password_hash::PasswordHasher` will be
unavailable (`password-hash` is only pulled `simple` feature is used).

It did not show up in CI, because `password-hash` is also provided as a
`dev-dependency` and is available when `cargo test` evaluates the
documentation.

Fixup #602