yescrypt: remove no-op `yescrypt_kdf_body` call (#647)
The original C code has it write into a buffer which isn't used.

Per the deleted comment this is apparently for demonstrating
implementation parity with the optimized implementation, which does some
form of preallocation which is unused in the reference implementation.

Additionally this converts `yescrypt_kdf_body` to returning a `Result`.