yescrypt: allocate state as a `Vec` (#663)
Notably this is the last usages of `malloc`/`free` in the codebase.

Also adds a lifetime to `PxwformCtx` and changes the constructor
argument to be a mutable slice.

It's still pointers internally, but this encapsulates a lot of the
unsafety, and allows us to remove `unsafe` from the constructor as it's
now (externally) safe to call.