yescrypt: add `Params::encode`
Initial support for encoding parameters for MCF strings.

This adapts the reference C code for encoding params from the
`yescrypt_encode_params_r` function. It currently optimizes for
closeness to the original C over making the code more idiomatic.

Notably the portions of the function for adding the `$y$` prefix and
encoding the salt as Base64 have been removed, as the function is
directly impl'd on the `Params` type and only deals with encoding those.

Params are encoded as a sequence of "LEB"-like variable-width integers.
In the original source code, encoding of the integers and the encoding
of Base64 are somewhat entangled, which is preserved here.

In the future, this can be made into an internal function, called by a
fuller implementation of the Modular Crypt Format password hashes used
by `yescrypt` (having been inherited from `scrypt`).

A handful of arbitrary test vectors generated against the reference C
`yescrypt_encode_params_r` have been added.