yescrypt: extract `util` module; bring MSRV down to 1.86 (#672)
Ideally we could ship a crate with MSRV 1.85 so we have a release that
works on stable Debian.

This vendors the code for `[T]::as_chunks_mut` and in the process
extracts a `util` module to contain the various shared free functions.

There's still a usage of `get_disjoint_mut` which keeps the MSRV at 1.86
which is a bit less trivial to just vendor like this, which is not
addressed for now, but ideally we'd have a temporary replacement for
that as well.

cc @conradludgate