yescrypt: "simple" API (#673)
Adds a new crate feature `simple` (on-by-default) which implements
initial support for an API which computes password hash strings in
Modular Crypt Format (MCF), with a leading `$y$` prefix.

This adapts the Base64 encoding from the yescrypt reference
implementation, as for whatever reason I couldn't get it to work with
`base64ct` (via the `mcf` crate).