yescrypt: fix miscomputed outputs
There were a few regressions introduced in the recent translation which
were not caught by the KATs or review.

The first was in 763d6f6 (mea culpa), which rewrote `wrapping_*`
arithmetic using standard infix operators which apply checked arithmetic
in debug. Notably in this commit a `wrapped_mul` was incorrectly
translated to a `/`.

The next is this logic gated a prehashing step, support for which was
removed in subsequent commits but not caught in part because the KATs
don't cover values of `N` that large (ironically, including the
recommended value). Such support has now been added back.

With that, we now pass both the `simple` password hash vector which was
failing before, and also the regression noted in #616 which has been
captured as a regression test.