sha-crypt: switch from `rand` to `rand_core`
Replaces the use of `Distribution` by first filling a buffer with random
bytes, then encoding it as Base64.

It seems the Base64 encoding is directly consumed by the algorithm, or
otherwise it would probably make sense to convert all usages of `salt`
to be raw bytes. That warrants further investigation.