Leverage `password-hash/getrandom` (#769)
Companion PR to RustCrypto/traits#2123

The aforementioned traits PR renames `PasswordHash::hash_password` to
`hash_password_with_salt`, so this PR updates the method name change
accordingly.

When the `getrandom` feature is enabled, it provides a new
`PasswordHash::hash_password` which *just* accepts a password as an
argument, relying on `getrandom` to internally generate a salt.

This updates all the code examples to use the new API so users don't
have to deal with randomly generating a salt.