sha-crypt: impl the `password-hash` crate traits
Rewrites the `simple` feature to use the `(Customized)PasswordHasher`
and `PasswordVerifier` traits, currently impl'd for a ZST.

The implementations are currently duplicated between SHA-256 and
SHA-512, but this PR leaves addressing that duplication for a followup
and just gets everything working to begin with.

The majority of the original test vectors are preserved.

The `CheckError` and `DecodeError` types have been replaced with
`password_hash::Error` and `mcf::Error` respectively.