sha-crypt: unify `ShaCrypt` trait impls for Sha256/512 (#773)
Adds a `ShaCryptCore` trait and impls it for `ShaCrypt<Sha256>` and
`ShaCrypt<Sha512>`, which specifies the MCF ID and core function,
including handling the digest-specific transposition tables.

This allows the `(Customized)PasswordHasher` and `PasswordVerifier`
trait impls to be generic over the digest function.

The `ShaCryptCore` is made public just in case anyone cares about
writing downstream generic code with it, however since the trait and
type it's impl'd for are both defined in `sha-crypt`, there's no way to
misuse foreign impls in the context of the crate, so sealing it is not
needed.