Rename `simple` features to `password-hash` (#776)
To reflect the feature activates the integration with the
`password-hash`, this renames the various `simple` features of the
crates in this repo to `password-hash` instead.

It also takes care of wiring up `rand_core` better.

The respective `simple` modules have either been named to `mcf` or `phc`
depending on which format is implemented.

This will make it possible to add another module to crates that support
both, e.g. `pbkdf2` and `scrypt` could both have a `mcf` module.

Closes #764